requirement.txt
Requirements:
python 3.12 installed
Pycharm Community Edition
Git

>>Install Django ,djangorestframework,djangorestframework-simplejwt
Using : pip install <package name>
>>to create a project
django-admin startproject projectname
>>to create an app,navigate to the project folder
django-admin startapp appname
>>configure the settings.py with the newly added appname, rest_framework and rest_framework_simplejwt
REST_FRAMEWORK = {
    'DEFAULT_AUTHENTICATION_CLASSES': (
        'rest_framework_simplejwt.authentication.JWTAuthentication',
    ),
}
>>
